HS-255 serializacja klas typu dao
