+++ core/src/main/java/pl/agh/edu/engine/attraction/AttractionSize.java
@@ -1,7 +1,13 @@
 package pl.agh.edu.engine.attraction;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public enum AttractionSize {
 	SMALL,
 	MEDIUM,
-	LARGE
+	LARGE;
+
+	static {
+		KryoConfig.kryo.register(AttractionSize.class);
+	}
 }
