+++ core/src/main/java/pl/agh/edu/engine/attraction/AttractionState.java
@@ -1,10 +1,16 @@
 package pl.agh.edu.engine.attraction;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public enum AttractionState {
 	ACTIVE,
 	BEING_BUILD,
 	CHANGING_SIZE,
 	SHUTTING_DOWN,
 	INACTIVE,
-	OPENING
+	OPENING;
+
+	static {
+		KryoConfig.kryo.register(AttractionState.class);
+	}
 }
