+++ core/src/main/java/pl/agh/edu/engine/attraction/AttractionType.java
@@ -1,7 +1,13 @@
 package pl.agh.edu.engine.attraction;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public enum AttractionType {
 	RESTAURANT,
 	SWIMMING_POOL,
-	SPA
+	SPA;
+
+	static {
+		KryoConfig.kryo.register(AttractionType.class);
+	}
 }
