+++ core/src/main/java/pl/agh/edu/engine/bank/TransactionType.java
@@ -1,6 +1,13 @@
 package pl.agh.edu.engine.bank;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public enum TransactionType {
 	INCOME,
-	EXPENSE
+	EXPENSE;
+
+	static {
+		KryoConfig.kryo.register(TransactionType.class);
+
+	}
 }
