+++ core/src/main/java/pl/agh/edu/engine/calendar/CalendarEvent.java
@@ -2,9 +2,13 @@ package pl.agh.edu.engine.calendar;
 
 import java.time.LocalDate;
 
+import pl.agh.edu.serialization.KryoConfig;
 import pl.agh.edu.utils.LanguageString;
 
 public record CalendarEvent(LocalDate date, LanguageString title, LanguageString description) implements Comparable<CalendarEvent> {
+    static {
+        KryoConfig.kryo.register(CalendarEvent.class);
+    }
 
     @Override
     public int compareTo(CalendarEvent o) {
