+++ core/src/main/java/pl/agh/edu/engine/client/Arrival.java
@@ -2,7 +2,13 @@ package pl.agh.edu.engine.client;
 
 import java.time.LocalTime;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public record Arrival(LocalTime time, ClientGroup clientGroup) implements Comparable<Arrival> {
+
+    static {
+        KryoConfig.kryo.register(Arrival.class);
+    }
 
     @Override
     public int compareTo(Arrival o) {
