+++ core/src/main/java/pl/agh/edu/engine/client/ClientGroup.java
@@ -4,9 +4,15 @@ import java.math.BigDecimal;
 import java.time.Duration;
 import java.util.List;
 
+import com.esotericsoftware.kryo.Kryo;
+import com.esotericsoftware.kryo.Serializer;
+import com.esotericsoftware.kryo.io.Input;
+import com.esotericsoftware.kryo.io.Output;
+
 import pl.agh.edu.engine.hotel.HotelVisitPurpose;
 import pl.agh.edu.engine.opinion.Opinion;
 import pl.agh.edu.engine.room.RoomRank;
+import pl.agh.edu.serialization.KryoConfig;
 
 public class ClientGroup {
 	private final List<Client> members;
@@ -16,6 +22,36 @@ public class ClientGroup {
 	private final Duration maxWaitingTime;
 	private final int numberOfNights;
 	public final Opinion opinion;
+
+	static {
+		KryoConfig.kryo.register(ClientGroup.class, new Serializer<ClientGroup>() {
+			@Override
+			public void write(Kryo kryo, Output output, ClientGroup object) {
+				kryo.writeObject(output, object.members, KryoConfig.listSerializer(Client.class));
+				kryo.writeObject(output, object.hotelVisitPurpose);
+				kryo.writeObject(output, object.desiredPricePerNight);
+				kryo.writeObject(output, object.desiredRoomRank);
+				kryo.writeObject(output, object.maxWaitingTime);
+				kryo.writeObject(output, object.numberOfNights);
+				kryo.writeObject(output, object.opinion);
+			}
+
+			@Override
+			public ClientGroup read(Kryo kryo, Input input, Class<? extends ClientGroup> type) {
+				ClientGroup clientGroup = new ClientGroup.Builder()
+						.members(kryo.readObject(input, List.class, KryoConfig.listSerializer(Client.class)))
+						.hotelVisitPurpose(kryo.readObject(input, HotelVisitPurpose.class))
+						.desiredPricePerNight(kryo.readObject(input, BigDecimal.class))
+						.desiredRoomRank(kryo.readObject(input, RoomRank.class))
+						.maxWaitingTime(kryo.readObject(input, Duration.class))
+						.numberOfNights(kryo.readObject(input, Integer.class))
+						.build();
+
+				KryoConfig.setPrivateFieldValue(clientGroup, "opinion", kryo.readObject(input, Opinion.class));
+				return clientGroup;
+			}
+		});
+	}
 
 	private ClientGroup(Builder builder) {
 		this.hotelVisitPurpose = builder.hotelVisitPurpose;
