+++ core/src/main/java/pl/agh/edu/engine/client/Gender.java
@@ -1,7 +1,13 @@
 package pl.agh.edu.engine.client;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public enum Gender {
 	MALE,
 	FEMALE,
-	OTHER
+	OTHER;
+
+	static {
+		KryoConfig.kryo.register(Gender.class);
+	}
 }
