+++ core/src/main/java/pl/agh/edu/engine/employee/Profession.java
@@ -1,7 +1,13 @@
 package pl.agh.edu.engine.employee;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public enum Profession {
 	CLEANER,
 	TECHNICIAN,
-	RECEPTIONIST
+	RECEPTIONIST;
+
+	static {
+		KryoConfig.kryo.register(Profession.class);
+	}
 }
