+++ core/src/main/java/pl/agh/edu/engine/employee/contract/TypeOfContract.java
@@ -1,5 +1,11 @@
 package pl.agh.edu.engine.employee.contract;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public enum TypeOfContract {
-	PERMANENT,
+	PERMANENT;
+
+	static {
+		KryoConfig.kryo.register(TypeOfContract.class);
+	}
 }
