+++ core/src/main/java/pl/agh/edu/engine/hotel/HotelType.java
@@ -1,7 +1,13 @@
 package pl.agh.edu.engine.hotel;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public enum HotelType {
 	HOTEL,
 	RESORT,
-	SANATORIUM
+	SANATORIUM;
+
+	static {
+		KryoConfig.kryo.register(HotelType.class);
+	}
 }
