+++ core/src/main/java/pl/agh/edu/engine/hotel/HotelVisitPurpose.java
@@ -1,7 +1,13 @@
 package pl.agh.edu.engine.hotel;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public enum HotelVisitPurpose {
 	VACATION,
 	BUSINESS_TRIP,
-	REHABILITATION
+	REHABILITATION;
+
+	static {
+		KryoConfig.kryo.register(HotelVisitPurpose.class);
+	}
 }
