+++ core/src/main/java/pl/agh/edu/engine/hotel/dificulty/DifficultyLevel.java
@@ -1,8 +1,14 @@
 package pl.agh.edu.engine.hotel.dificulty;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public enum DifficultyLevel {
 	EASY,
 	MEDIUM,
 	HARD,
-	BRUTAL
+	BRUTAL;
+
+	static {
+		KryoConfig.kryo.register(DifficultyLevel.class);
+	}
 }
