+++ core/src/main/java/pl/agh/edu/engine/opinion/bucket/RoomBreakingOpinionBucket.java
@@ -3,10 +3,41 @@ package pl.agh.edu.engine.opinion.bucket;
 import java.util.Objects;
 import java.util.Optional;
 
+import com.esotericsoftware.kryo.Kryo;
+import com.esotericsoftware.kryo.Serializer;
+import com.esotericsoftware.kryo.io.Input;
+import com.esotericsoftware.kryo.io.Output;
+
+import pl.agh.edu.serialization.KryoConfig;
+
 public class RoomBreakingOpinionBucket extends OpinionBucket {
 	private boolean gotBrokenRoom = false;
 	private boolean roomBroke = false;
 	private boolean repaired = false;
+
+	static {
+		KryoConfig.kryo.register(RoomBreakingOpinionBucket.class, new Serializer<RoomBreakingOpinionBucket>() {
+			@Override
+			public void write(Kryo kryo, Output output, RoomBreakingOpinionBucket object) {
+				kryo.writeObject(output, object.weight);
+				kryo.writeObject(output, object.gotBrokenRoom);
+				kryo.writeObject(output, object.roomBroke);
+				kryo.writeObject(output, object.repaired);
+
+			}
+
+			@Override
+			public RoomBreakingOpinionBucket read(Kryo kryo, Input input, Class<? extends RoomBreakingOpinionBucket> type) {
+				RoomBreakingOpinionBucket roomBreakingOpinionBucket = new RoomBreakingOpinionBucket(kryo.readObject(input, Integer.class));
+
+				roomBreakingOpinionBucket.gotBrokenRoom = kryo.readObject(input, Boolean.class);
+				roomBreakingOpinionBucket.roomBroke = kryo.readObject(input, Boolean.class);
+				roomBreakingOpinionBucket.repaired = kryo.readObject(input, Boolean.class);
+
+				return roomBreakingOpinionBucket;
+			}
+		});
+	}
 
 	public RoomBreakingOpinionBucket(int weight) {
 		super(weight);
