+++ core/src/main/java/pl/agh/edu/engine/opinion/bucket/RoomCleaningOpinionBucket.java
@@ -3,10 +3,41 @@ package pl.agh.edu.engine.opinion.bucket;
 import java.util.Objects;
 import java.util.Optional;
 
+import com.esotericsoftware.kryo.Kryo;
+import com.esotericsoftware.kryo.Serializer;
+import com.esotericsoftware.kryo.io.Input;
+import com.esotericsoftware.kryo.io.Output;
+
+import pl.agh.edu.serialization.KryoConfig;
+
 public class RoomCleaningOpinionBucket extends OpinionBucket {
 	private boolean gotCleanRoom = false;
 	private int cleanRoomCounter = 0;
 	private final int numberOfNights;
+
+	static {
+		KryoConfig.kryo.register(RoomCleaningOpinionBucket.class, new Serializer<RoomCleaningOpinionBucket>() {
+			@Override
+			public void write(Kryo kryo, Output output, RoomCleaningOpinionBucket object) {
+				kryo.writeObject(output, object.weight);
+				kryo.writeObject(output, object.numberOfNights);
+				kryo.writeObject(output, object.gotCleanRoom);
+				kryo.writeObject(output, object.cleanRoomCounter);
+
+			}
+
+			@Override
+			public RoomCleaningOpinionBucket read(Kryo kryo, Input input, Class<? extends RoomCleaningOpinionBucket> type) {
+				RoomCleaningOpinionBucket roomCleaningOpinionBucket = new RoomCleaningOpinionBucket(
+						kryo.readObject(input, Integer.class),
+						kryo.readObject(input, Integer.class));
+				roomCleaningOpinionBucket.gotCleanRoom = kryo.readObject(input, Boolean.class);
+				roomCleaningOpinionBucket.cleanRoomCounter = kryo.readObject(input, Integer.class);
+
+				return roomCleaningOpinionBucket;
+			}
+		});
+	}
 
 	public RoomCleaningOpinionBucket(int weight, int numberOfNights) {
 		super(weight);
