+++ core/src/main/java/pl/agh/edu/engine/room/RoomRank.java
@@ -1,7 +1,13 @@
 package pl.agh.edu.engine.room;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public enum RoomRank {
 	ECONOMIC,
 	STANDARD,
-	DELUXE
+	DELUXE;
+
+	static {
+		KryoConfig.kryo.register(RoomRank.class);
+	}
 }
