+++ core/src/main/java/pl/agh/edu/utils/Pair.java
@@ -1,7 +1,13 @@
 package pl.agh.edu.utils;
 
+import pl.agh.edu.serialization.KryoConfig;
+
 public record Pair<F, S>(F first, S second) {
     public static <F, S> Pair<F, S> of(F first, S second) {
         return new Pair<>(first, second);
+    }
+
+    static {
+        KryoConfig.kryo.register(Pair.class);
     }
 }
